# openapi.yaml
openapi: 3.0.3
info:
  title: Alimenti API
  description: >
    API REST per la gestione di alimenti (composizione CREA/INRAN), tabelle
    di riferimento LARN, ricette, pazienti con piani alimentari e
    appuntamenti, piramide Smartfood. Trasformazione in API REST
    dell'applicativo "Registro Alimenti" (Project Work — Sviluppo di
    un'applicazione full-stack API-based per un'organizzazione del settore
    sanitario, CdS Informatica per le Aziende Digitali).

    Autenticazione: un singolo utente amministratore. Le operazioni di sola
    lettura su alimenti, LARN, ricette e smartfood sono pubbliche; le
    operazioni di scrittura su questi moduli e TUTTE le operazioni sui
    pazienti (dati clinici) richiedono il login (vedi /auth/login). La
    sessione e' mantenuta con un cookie firmato.
  version: "1.0.0"
servers:
  - url: /api
    description: Stesso host del front-end (modalita' consigliata)
...
security:
  - cookieAuth: []
paths:
...
  /alimenti:
    get:
      tags: [Alimenti]
      summary: Elenco paginato degli alimenti
      security: []
      parameters:
        - $ref: "#/components/parameters/q"
        - name: categoria
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/page"
      responses:
        "200":
          description: Elenco alimenti
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginaMeta"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/AlimentoSintetico" }
    post:
      tags: [Alimenti]
      summary: Crea un alimento
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AlimentoInput" }
      responses:
        "201":
          description: Alimento creato
          content:
            application/json: { schema: { $ref: "#/components/schemas/Alimento" } }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
...
  /alimenti/{id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      tags: [Alimenti]
      summary: Dettaglio alimento
      security: []
      responses:
        "200":
          description: Alimento
          content:
            application/json: { schema: { $ref: "#/components/schemas/Alimento" } }
        "404": { $ref: "#/components/responses/NonTrovato" }
    put:
      tags: [Alimenti]
      summary: Aggiorna un alimento
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AlimentoInput" }
      responses:
        "200":
          description: Alimento aggiornato
          content:
            application/json: { schema: { $ref: "#/components/schemas/Alimento" } }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
    delete:
      tags: [Alimenti]
      summary: Elimina un alimento
      responses:
        "204": { description: Eliminato }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }

  # ------------------------------------------------------------------
  # LARN (generico su 15 tabelle)
  # ------------------------------------------------------------------
...
  /larn/{table_key}:
    parameters:
      - $ref: "#/components/parameters/tableKeyPath"
    get:
      tags: [LARN]
      summary: Elenco righe di una tabella LARN (con filtri)
      security: []
      responses:
        "200": { description: Elenco righe }
        "404": { $ref: "#/components/responses/NonTrovato" }
    post:
      tags: [LARN]
      summary: Crea una riga nella tabella LARN indicata
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, description: "Campi specifici della tabella, vedi GET /larn/tables" }
      responses:
        "201": { description: Riga creata }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
  /larn/{table_key}/{id}:
    parameters:
      - $ref: "#/components/parameters/tableKeyPath"
      - $ref: "#/components/parameters/idPath"
    get:
      tags: [LARN]
      summary: Dettaglio riga
      security: []
      responses:
        "200": { description: Riga }
        "404": { $ref: "#/components/responses/NonTrovato" }
    put:
      tags: [LARN]
      summary: Aggiorna una riga
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: Riga aggiornata }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
    delete:
      tags: [LARN]
      summary: Elimina una riga
      responses:
        "204": { description: Eliminata }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }

  # ------------------------------------------------------------------
  # Ricette
  # ------------------------------------------------------------------
...
  /pazienti/{id}/piani/{piano_id}/giorni/{giorno}/alimenti:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - name: piano_id
        in: path
        required: true
        schema: { type: integer }
      - $ref: "#/components/parameters/giornoPath"
    post:
      tags: [Piani alimentari]
      summary: Collega un alimento (con quantita' in grammi) a un pasto del giorno
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/AlimentoPastoInput" } }
      responses:
        "201": { description: Alimento collegato }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
  /pazienti/{id}/piani/{piano_id}/giorni/{giorno}/alimenti/{link_id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - name: piano_id
        in: path
        required: true
        schema: { type: integer }
      - $ref: "#/components/parameters/giornoPath"
      - name: link_id
        in: path
        required: true
        schema: { type: integer }
    delete:
      tags: [Piani alimentari]
      summary: Rimuove un alimento collegato a un pasto
      responses:
        "204": { description: Rimosso }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }

  # ------------------------------------------------------------------
  # Appuntamenti
  # ------------------------------------------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Cookie di sessione impostato da POST /auth/login.
  parameters:
    q:
      name: q
      in: query
      description: Ricerca libera
      schema: { type: string }
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    idPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
    tableKeyPath:
      name: table_key
      in: path
      required: true
      description: 'Chiave di una delle 15 tabelle LARN (vedi GET /larn/tables), es. "proteine", "acqua", "vitamine_pri_ai".'
      schema: { type: string }
    giornoPath:
      name: giorno
      in: path
      required: true
      schema:
        type: string
        enum: [Lunedi, Martedi, Mercoledi, Giovedi, Venerdi, Sabato, Domenica]
  responses:
    NonTrovato:
      description: Risorsa non trovata
      content:
        application/json: { schema: { $ref: "#/components/schemas/Errore" } }
    NonAutenticato:
      description: Login richiesto
      content:
        application/json: { schema: { $ref: "#/components/schemas/Errore" } }
    DatiNonValidi:
      description: Dati non validi
      content:
        application/json: { schema: { $ref: "#/components/schemas/ErroreValidazione" } }
  schemas:
    Errore:
      type: object
      properties:
        error: { type: string }
    ErroreValidazione:
      type: object
      properties:
        error: { type: string }
        details:
          type: array
          items: { type: string }
    PaginaMeta:
      type: object
      properties:
        page: { type: integer }
        total_pages: { type: integer }
        total: { type: integer }
        page_size: { type: integer }
    AlimentoSintetico:
      type: object
      description: Colonne principali, usate nell'elenco paginato
      properties:
        id: { type: integer }
        codice_alimento: { type: string }
        nome_alimento: { type: string }
        categoria: { type: string }
        energia_kcal: { type: number }
        proteine_g: { type: number }
        lipidi_g: { type: number }
        carboidrati_disponibili_g: { type: number }
    Alimento:
      type: object
      description: >
        Scheda completa (~50 campi nutrizionali per 100 g di parte edibile:
        energia, macronutrienti, vitamine, minerali). Vedi GET /alimenti/campi
        per l'elenco completo con etichette e unita' di misura.
      properties:
        id: { type: integer }
        codice_alimento: { type: string }
        nome_alimento: { type: string }
        categoria: { type: string }
        energia_kcal: { type: number }
        proteine_g: { type: number }
        lipidi_g: { type: number }
        carboidrati_disponibili_g: { type: number }
        fibra_totale_g: { type: number }
        sodio_mg: { type: number }
      additionalProperties: true
    AlimentoInput:
      allOf:
        - $ref: "#/components/schemas/Alimento"
      description: Stessi campi di Alimento (senza "id", assegnato dal server).
...
    AlimentoPastoInput:
      type: object
      required: [pasto, codice_alimento, quantita_g]
      properties:
        pasto: { type: string, enum: [colazione, spuntino_mattina, pranzo, spuntino_pomeriggio, cena] }
        codice_alimento: { type: string }
        quantita_g: { type: number, exclusiveMinimum: 0 }
        note: { type: string }
...